requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9

# AWS SDK
boto3>=1.35.0
botocore>=1.35.0

# ML / Model
torch>=2.9.0
transformers>=4.44.0
peft>=0.12.0
Pillow>=10.4.0
timm>=1.0.8
einops>=0.8.0

# Dataset
datasets>=2.21.0
kaggle>=1.6.17

# Sarvam API
requests==2.32.0
httpx==0.27.0

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1

# Monitoring & Logging
structlog==24.4.0
watchtower==3.2.0  # CloudWatch log handler

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0  # For async test client
moto[s3,dynamodb,sagemaker]==5.0.0  # AWS mocking

# Code Quality
ruff==0.6.0
mypy==1.11.0

# Frontend (demo)
streamlit==1.38.0

# Training extras
scikit-learn>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0
pandas>=2.2.0
numpy>=1.26.0

# Deployment
gunicorn==23.0.0
docker==7.1.0
